Migrate APIs into C API and set up unit test framework.
